Move unit test to a new file for manual tests.
